Asset Protection/requirements.txt
opencv-python>=4.8.0
imagehash>=4.3.1
Pillow>=10.0.0
ultralytics>=8.0.0
numpy>=1.24.0
flask>=3.0.0
flask-cors>=4.0.0
yt-dlp>=2023.12.30
